# Optimize llvm-nm