--- a/tools/building.py
+++ b/tools/building.py
@@ -389,10 +389,10 @@ def llvm_nm_multiple(files):
       if nl < 0:
         break
       colon = results.rfind(':', i, nl)
-      if colon >= 0 and results[colon+1] == '\n': # New file start?
+      if colon >= 0 and results[colon + 1] == '\n': # New file start?
         nm_cache[filename] = parse_symbols(results[file_start:i-1])
         filename = results[i:colon].strip()
-        file_start = colon+2
+        file_start = colon + 2
       i = nl + 1
 
     nm_cache[filename] = parse_symbols(results[file_start:])