@kripken 2021-02-10
We have a singleton multiprocessing pool, so if it's created anyhow I'm not sure this matter? Other uses are the JS optimizer (in wasm2js, but not wasm) and system library building, but there might be more.

If this is helpful to most builds, then this makes sense. But then can we do it as a wrapper around the pool, so it helps the other cases as well, and not just llvm-nm specifically? I assume the same reasoning applies there too.