@juj 2021-02-11
Actually I now realized that we don't need the multiprocessing pool at all for this task. It is bad, slow, buggy, clunky, bloated, and all the other FUD adjectives that one can associate with a piece of software code they don't like. A major source of headache. :(

But now when looking at the problem this morning, I realize that the good old `subprocess.Popen()` is already an asynchronous parallel API - let's just use that. Rewrote the PR to do parallel `Popen()` calls instead. Much simpler and shorter.

Here are benchmarks in tiny and large contexts:

A tiny sized project
============

https://github.com/juj/wasm_webgpu , consists of a single .cpp file that builds to a .a library, and then three individual main apps that each have a single .cpp file, and each link in the static library. Generates two llvm-nm calls per app, for a total of 6x llvm-nm.

Before this PR, a debug iteration build took 3.3 seconds, of that time llvm-nm code eats up 2.3 seconds:

http://clb.confined.space/dump/toolchain_profiler.results_20210211_1209.html

![tiny_baseline](https://user-images.githubusercontent.com/225351/107627988-b12f5580-6c68-11eb-9101-5d917820ac56.png)

After this PR, a debug iteration build takes 650 msecs, of that time llvm-nm code is 14 msecs: http://clb.confined.space/dump/toolchain_profiler.results_20210211_1214.html

![tiny_optimized](https://user-images.githubusercontent.com/225351/107628002-b5f40980-6c68-11eb-99df-b86684c37bff.png)

A large sized project
=============

https://github.com/gabrielcuvillier/d3wasm/ , Gabriel Cuvillier's port of Doom 3 over to Wasm. Final link consists of 225 calls to llvm-nm. CC @gabrielcuvillier 

Before this PR, a release link step took 31.7 seconds, of which llvm-nm calls took up 3.8 seconds:

http://clb.confined.space/dump/d3wasm_baseline.html

![d3wasm_baseline](https://user-images.githubusercontent.com/225351/107628045-c1473500-6c68-11eb-88a5-1bb37805ceab.png)

After this PR, a release link step takes 28.9 seconds, of which llvm-nm calls take 900 msecs:
 
http://clb.confined.space/dump/d3wasm_llvm_nm_optimized.html

![d3wasm_optimized](https://user-images.githubusercontent.com/225351/107628058-c6a47f80-6c68-11eb-9836-6367c49ac746.png)

This should be a win across the board.

I dream of a world where python multiprocessing is not used at all.