@juj 2021-02-11
Actually, it looks like `llvm-nm` supports taking in a list of files. Doing that, 3dwasm build looks simply like

http://clb.confined.space/dump/d3wasm_optimized_2.html

![d3wasm_optimized_2](https://user-images.githubusercontent.com/225351/107658667-07af8a80-6c8f-11eb-8d92-16c78e343595.png)

taking up 570ms for the `llvm-nm` processing, a further -37% time reduction from the above, and a -85% time reduction to baseline.