@juj 2021-02-13
Ran through the profile nevertheless (btw would you have a chance to review #13464 at some point? I split it separate from this to avoid review churn, but it is awkward to keep octopus merging back working branches together after having posted a PR)

Result is the same 11ms from before. Compiler sanity checks and compiler.js dominate.

![tiny_optimized_2](https://user-images.githubusercontent.com/225351/107844792-3f2d4c80-6ddf-11eb-8de8-fa6bf4b08138.png)