@sbc100 2021-03-15
Looks like this broke windows users with many input files: https://github.com/emscripten-core/emscripten/pull/13465